still working on it

+++ lru_cache.h
@@ -1,7 +1,6 @@
 #ifndef lru_cache_H
 #define lru_cache_H
 
-
-unsigned long long int LRUCaching(int cacheSize, unsigned long long int randomNum);
+int LRUCaching(int cacheSize, unsigned long long int randomNum);
 
 #endif
